system: keep track of /etc
Add a script to compare /etc to /usr/etc, along with a directory
specifying a list of changes that should have been made.

Two special forms of file are supported for the changes directory:

  - a symlink with the target "ignore" represents a file or subdirectory
    which should be ignored

  - a symlink with the target "delete" represents a file present in
    /usr/etc which should be deleted in /etc

Otherwise, any file or symlink present in the changes directory is
expected to appear in /etc in the same way.  If a file is not present in
the changes directory, then /etc is expected to have the same content as
/usr/etc.

There's no support for actually modifying /etc — this is just a checking
tool.