user: rename 'stow' to 'stash', and install it
Move `stow` into a package inside `user/`, renaming it to `stash` and
installing it in the PATH.

Add a symlink from `user/stash`, where we use it from installing the content of
`user/packages`.  It didn't really make sense to have this at the toplevel of
the repository, since `system/` never used it.